fix: Encode SD-JWT disclosure value without padding
The SD-JWT specification uses base64url to encode the disclosure. As per section 1.2 (Conventions and Terminology), this
refers to URL-safe Base64 without padding. Kotlins Base64.UrlSafe encoder, however uses padding. Depending on the input,
this leads to padding characters in output, which is against the specification and also fails in the SDJwt.parse method
due to the regex check.

Signed-off-by: Tobias Wich <tobias.wich@ecsec.de>