refactor: use access token instead of token response in offer processor
In the context of OID4VCI flows, an access token is always included in the token response. Added missing validation about whether the call was successful and to check whether the access token is included in the response as it should. Propagated the acquired access token all the way to the response of the respective handler and marked it as optional since MS Entra does require it. Test cases have been refactored accordingly as well.